lnew ineHeaderTableRows component, removed lineHeader copy button that isnt needed
