+++ src/components/sidebar/lineView/lineHeaderTableRows.scss
@@ -1,0 +1,53 @@
+@import '../../shared/styles/common';
+@import '../../shared/styles/flexbox';
+@import '../../shared/styles/form';
+
+$cellPadding: 5px 2px 5px 2px; // same as in lineHeaderTable.scss
+
+.lineHeaderTableRow {
+    font-size: $smallFontSize;
+    border-bottom: 1px solid #dfdfdf;
+
+    .lineHeaderTableNameCell {
+        max-width: 180px;
+        padding: 5px 2px 5px 2px;
+    }
+
+    .lineHeaderTableCalendarCell {
+        min-width: 125px;
+        padding: $cellPadding;
+        vertical-align: top;
+
+        .timeInput {
+            padding-top: 0px;
+        }
+    }
+
+    .lineHeaderTableButtonCell {
+        vertical-align: top;
+        padding: $cellPadding;
+    }
+
+    .lineHeaderButton {
+        display: flex;
+        font-size: $largeFontSize;
+        padding: 5px;
+        margin-top: 4px;
+        margin-bottom: auto;
+        border-radius: 5px;
+    }
+
+    .disabledRemoveLineHeaderButton {
+        color: $disabledGrey;
+        background-color: $backgroundlightGrey;
+    }
+
+    .removeLineHeaderButton:hover {
+        color: white;
+        background-color: $red;
+    }
+}
+
+.lineHeaderRowHighlight {
+    background-color: $lightblue;
+}
